refactor(sonar): lift isShiftWindowFull into usePlanningGrid to remove duplication
The window-day-capacity gate (`timeWindowById` + `isShiftWindowFull` derived from
`getRemainingQuota`) was identical in `planning-week-view` and `day/day-grid` —
~18 new lines duplicated in both files (Sonar: 29.8% new-code duplication). Move
the lookup + callback into `use-planning-grid` and consume it from `planningGrid`
in both views. No behavior change.
